grpc_server/requirements.txt
arrow==0.17.0
click==7.1.2
Flask==1.1.2
grpcio==1.35.0
grpcio-tools==1.35.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
protobuf==3.15.1
python-dateutil==2.8.1
redis==3.5.3
rq==1.7.0
rq-dashboard==0.6.1
six==1.15.0
Werkzeug==1.0.1
